+++ examples/readme.ts
@@ -7,18 +7,18 @@ import fs from 'node:fs';
 await flow.registerFont(new URL('../assets/Roboto/Roboto-Regular.ttf', import.meta.url));
 await flow.registerFont(new URL('../assets/Roboto/Roboto-Bold.ttf', import.meta.url));
 
-// Always create styles at the top-level of your module if you can
-const divStyle: flow.DeclaredStyle = {
+// Always create styles at the top-level of your module if you can.
+const divStyle = flow.style({
   backgroundColor: {r: 28, g: 10, b: 0, a: 1},
   textAlign: 'center',
   color: {r: 179, g: 200, b: 144, a: 1}
-};
+});
 
-// Since we're creating styles directly, colors have to be defined numerically
-const spanStyle: flow.DeclaredStyle = {
+// Since we're creating styles directly, colors are numbers
+const spanStyle = flow.style({
   color: {r: 115, g: 169, b: 173, a: 1},
   fontWeight: 700
-};
+});
 
 // Create a DOM
 const rootElement = flow.dom(