+++ CHANGELOG.md
@@ -8,8 +8,13 @@ project adheres to [Semantic Versioning](http://semver.org/).
 (Unreleased)
 ==================
 ### Changed
+* Styles must now be passed through `flow.style` before being given to `h`.
+* `cascadeStyles` has been removed. Pass an array of styles to `h` instead.
+* Removed `getRootStyle`
+
 ### Added
 * Added support for the `zoom` property
+* Support for multiple styles on an element
 
 ### Fixed